requirements.txt
inflect==2.1.0
irc==17.0
jaraco.classes==2.0
jaraco.collections==2.0
jaraco.functools==2.0
jaraco.itertools==4.4.1
jaraco.logging==2.0
jaraco.stream==2.0
jaraco.text==2.0
more-itertools==5.0.0
python-dotenv==0.10.1
pytz==2018.9
six==1.12.0
tempora==1.14
